Create new controllers and fxml for Nurse
We changed the bypass login function in the login controller to the nurse dashbaoard for testing than back to doctor dashboard.
Change the controller for the new nurse fxml to nurse controllers and not the doctors (we used the doctor fxml as a template which included the doctor controllers).
Created new nurse controllers with the doctor template and rename the file to Nurse_ and the class name to match.
We found a bug in the Nurse Patient Info View fxml which directed to the Patient Information fxml and not the directory. This was resolved by creating a new function switch to dashbaord and changing the onclick() in the fxml. The doctor patient info view fxml had the same bug and was also resolved in this update.
Tested fine.
To Do:  Appointment/calendar