Fix Next Appointment on Dashboard
The next appointment was looping through the entry and the last appointment was displayed.
We removed the loops in the dashboard and nurse dashboard; we used the first index of the entry instead.
Tried to serialize the entry and failed on the nurse dashboard by implementing serializable to the Calendar App.