Added a calendar entries from the DB to the calendar
Updated the Dashboard and Nurse Dashboard with a query to the doctor calendar database and collected the entries from each row to be added to the calendar , since the calendar is empty when opened.
Cleaned up the patient Directory Controller and nurse patient directory
 controller to print useful information instead of numbers.

Tested both Dashboard controllers and worked fine.