Updated the controller to ensure that next appointment  was correct.
The next appointment was the last appointment after the local time.
This was fixed by comparing the next appointment with the new entry and ensure that the new entry is not more than the next appointment, remember the next appointment must be more than the local time.

We found a bug with the calendar fxml when changing the controller to the nurse dashboard controller. This was fixed by only setting the text if the next appointment is not null like other dashboard bug found earlier

Test nurse dashboard and dashboard controllers fine.