Clean up the code
Copied the stop function from Toby's branch.
Deleted all the old controller files.
Ensured that the User Session class works for obtain the name which is being used in the stop function in main.
Created a test for user session.
Login controller will use the User session to set the email address.
Tested fine and the app logged out the user.